+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "svfrontend",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
